Extract shared helpers into pyriemann/_helpers.py
Move check_function, check_weights, check_init, check_param_in_func,
is_square, is_real_type, and the deprecated decorator into a neutral
_helpers.py module at the package root with no internal pyriemann
dependencies.

Update utils/utils.py, utils/test.py, and utils/docs.py to re-export
from _helpers for backward compatibility.
